+++ configure.py
@@ -335,7 +335,6 @@ scylla_core = (['database.cc',
                  'sstables/compress.cc',
                  'sstables/row.cc',
                  'sstables/partition.cc',
-                 'sstables/filter.cc',
                  'sstables/compaction.cc',
                  'sstables/compaction_strategy.cc',
                  'sstables/compaction_manager.cc',
